% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_comb}
\alias{plot_comb}
\title{Create a Comparison Plot of Pollen Concentrations for different traps and lines}
\usage{
plot_comb(resolution, traps, rm_zeros)
}
\arguments{
\item{resolution}{What temporal resolution should be plotted c("daily", "12hour", "6hour", "3hour", "hourly")}

\item{traps}{Which traps should be plotted c(poleno2, poleno4, poleno5, hirst)}

\item{rm_zeros}{Should zero pollen measurements be removed from the plot}
}
\value{
A data frame with assets allocated to 5 main categories
}
\description{
Create a Comparison Plot of Pollen Concentrations for different traps and lines
}
