% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_hirst}
\alias{plot_hirst}
\title{Create a Comparison Plot of Pollen Concentrations for different traps and lines}
\usage{
plot_hirst(species, resolution, group, traps, rm_zeros, combined)
}
\arguments{
\item{species}{The selected Pollen Type}

\item{resolution}{What temporal resolution should be plotted c("daily", "12hour", "6hour", "3hour", "hourly")}

\item{group}{Either compare traps or lines within traps c("line", "trap")}

\item{traps}{Which traps should be plotted c(2, 4, 6)}

\item{rm_zeros}{Should zero pollen measurements be removed from the plot}

\item{combined}{Should the function return the combined plots with title or seperate plots}
}
\value{
A data frame with assets allocated to 5 main categories
}
\description{
Create a Comparison Plot of Pollen Concentrations for different traps and lines
}
