% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_hist_dt}
\alias{plot_hist_dt}
\title{Fit t distribution to relative differences from the mean and then create daily samples of of it and return a ggplot with the fitted densitiy function}
\usage{
plot_hist_dt(
  data,
  species,
  n_test = 1,
  samples = 10000,
  plots = FALSE,
  tdist = TRUE,
  y_label = 2
)
}
\arguments{
\item{data}{The dataframe containing the measurements}

\item{species}{The selected Pollen Type}

\item{n_test}{Number of times the cvm and ad tests should be carried out}

\item{samples}{Number of samples created for each day}

\item{plots}{If plots is TRUE, histogram and density plots are returned, if FALSE the fct returns sampled daily values}

\item{tdist}{Should a t distribution be fitted to the data?}

\item{ylabel}{Where on the y-axis should the label be plotted}
}
\value{
A list of ggplots
}
\description{
Fit t distribution to relative differences from the mean and then create daily samples of of it and return a ggplot with the fitted densitiy function
}
