% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pollen}
\alias{plot_pollen}
\title{Create a Comparison Plot of Pollen Concentrations for different traps and lines}
\usage{
plot_pollen(species, resolution, group, traps)
}
\arguments{
\item{species}{The selected Pollen Type}

\item{resolution}{What temporal resolution should be plotted c("daily", "12hour", "6hour", "3hour", "hourly")}

\item{group}{Either compare traps or lines within traps c("line", "trap")}

\item{traps}{Which traps should be plotted c(2, 4, 6)}
}
\value{
A data frame with assets allocated to 5 main categories
}
\description{
Create a Comparison Plot of Pollen Concentrations for different traps and lines
}
