% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_spi}
\alias{plot_spi}
\title{Plot boxplots for n spi samples of one pollen species}
\usage{
plot_spi(data, species, samples, n = 1000, xlim = c(0, 35000))
}
\arguments{
\item{data}{The dataframe containing the measurements}

\item{species}{The selected Pollen Type}

\item{samples}{The sampled daily values (created by plot_hist_dt())}

\item{n}{Number of samples drawn for each day}

\item{xlim}{Limits of the x-axis in the boxplots}
}
\value{
A list of ggplots
}
\description{
Plot boxplots for n spi samples of one pollen species
}
