% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_na}
\alias{set_na}
\title{Sets measurements to NA for all times, where no trap measured anything}
\usage{
set_na(data)
}
\arguments{
\item{data}{A list of dataframes for the measurements}
}
\value{
A list of dataframes with NA values
}
\description{
Sets measurements to NA for all times, where no trap measured anything
}
